Trying to add an image ...attempt 3
THe url should be the address of the image..it should end with jpeg or png.  PDF does not work..it shoots an error (non-image content) :)
Another thing learnt...I ediited the markdown file...did not do the commit change thing...liked the change and decided to add the commit changes...it wont let you do any commit changes unless to make a change to the markdown file.  I have three exclamation marks so that I could add the 'commit changes' thing