Create WEEK2assignment
Line 1: gives us how many total sequences are there.  Answer is 85
Line 2: Sends the list of sequences to a file named listofproteins_coffeechloroplastgenome.txt
Line 3: Kind of wraptexts the sequence.  If a single sequences continues on the next line(s), this command converts it into a single sequence per line.
Line 4: Counts the number of sequences that have the word 'photosystem' in their name.  Answer is 22.
Line 5: sends the sequence and the sequence name for the name that matches 'ATP synthase CF1 beta subunit' and sends it to a file named ATPCF1_inonelinecoffee_chloroplastgenome.fasta