grep versus awk!
Discussed answers with WIll and realized that the Photosystem subunits are 3 not 22.  Tried to pipe two grep commands...isnt working for me. 
Long method: grep one string and put it in a file then grep the other string in that file.  Violates the rule: KEEP THE CODE SHORT and simple.
awk works well with the string.