Update WEEK2assignment
Step 1: looks for '>' sign to determine the total number of sequences.  ANSWER IS 85.
Step 2: Makes a list of the total protein sequences and saves it as listofproteins_coffeechloroplastgenome.txt
Step 3: Works as a wrap text so that line-breaks in the sequence are removed.
Step 4: saves all the names with the strings 'photosystem' in them into a file named photosystemlist.fasta.
Step 5: counts all the names that has the string 'subunit' in it from the file photosystemlist.fasta
Step 7: using a single awk step to do Steps 4 and 5.  THe total number of sequences for 'photosystem subunit' are 3.
Step 8: saves names containing "ATP synthase CF1 beta subunit" and their respective sequences in a file named 'ATPCF1_inonelinecoffee_chloroplastgenome.fasta'.  THere is 1 sequence.